All site changes
* all bitrix update

+++ bitrix/modules/sale/lib/helpers/admin/blocks/orderbasketshipment.php
@@ -275,6 +275,7 @@ class OrderBasketShipment extends OrderBasket
 			$basketItem = $item->getBasketItem();
 			if ($basketItem)
 			{
+				$systemShipmentItem = null;
 				if ($systemShipmentItemCollection)
 				{
 					/** @var \Bitrix\Sale\ShipmentItemCollection $systemShipmentItemCollection */
@@ -283,18 +284,28 @@ class OrderBasketShipment extends OrderBasket
 
 				$productId = $basketItem->getProductId();
 
-				if ($basketItem->getField("MODULE") == "catalog" && !empty($catalogProductsFields[$productId]))
+				if ($basketItem->getField("MODULE") === "catalog" && !empty($catalogProductsFields[$productId]))
+				{
 					$params = $catalogProductsFields[$productId];
+				}
 
 				if (intval($basketItem->getField("MEASURE_CODE")) > 0)
+				{
 					$params["MEASURE_CODE"] = intval($basketItem->getField("MEASURE_CODE"));
+				}
 				elseif (!isset($params["MEASURE_CODE"]))
+				{
 					$params["MEASURE_CODE"] = 0;
+				}
 
-				if($basketItem->getField("MEASURE_NAME") <> '')
+				if (!empty($basketItem->getField("MEASURE_NAME")))
+				{
 					$params["MEASURE_TEXT"] = $basketItem->getField("MEASURE_NAME");
+				}
 				elseif(!isset($params["MEASURE_TEXT"]))
+				{
 					$params["MEASURE_TEXT"] = "";
+				}
 
 				if ($basketItem->isBundleParent())
 				{
@@ -310,23 +321,28 @@ class OrderBasketShipment extends OrderBasket
 				$params["PRODUCT_PROVIDER_CLASS"] = $basketItem->getProvider();
 				$params["NAME"] = $basketItem->getField("NAME");
 				$params["MODULE"] = $basketItem->getField("MODULE");
+				$params['TYPE'] = (int)$basketItem->getField('TYPE');
+				$params['BARCODE_INFO'] = [];
 
 				$itemStoreCollection = $item->getShipmentItemStoreCollection();
-
-				/** @var \Bitrix\Sale\ShipmentItemStore $barcode */
-				$params['BARCODE_INFO'] = array();
-				foreach ($itemStoreCollection as $barcode)
+				if ($itemStoreCollection)
 				{
-					$storeId = $barcode->getStoreId();
-					if (!isset($params['BARCODE_INFO'][$storeId]))
-						$params['BARCODE_INFO'][$storeId] = array();
+					/** @var \Bitrix\Sale\ShipmentItemStore $barcode */
+					foreach ($itemStoreCollection as $barcode)
+					{
+						$storeId = $barcode->getStoreId();
+						if (!isset($params['BARCODE_INFO'][$storeId]))
+						{
+							$params['BARCODE_INFO'][$storeId] = [];
+						}
 
-					$params['BARCODE_INFO'][$storeId][] = array(
-						'ID' => $barcode->getId(),
-						'BARCODE' => $barcode->getBarcode(),
-						'MARKING_CODE' => $barcode->getMarkingCode(),
-						'QUANTITY' => $barcode->getQuantity()
-					);
+						$params['BARCODE_INFO'][$storeId][] = [
+							'ID' => $barcode->getId(),
+							'BARCODE' => $barcode->getBarcode(),
+							'MARKING_CODE' => $barcode->getMarkingCode(),
+							'QUANTITY' => $barcode->getQuantity(),
+						];
+					}
 				}
 
 				if (!$shipment->isSystem())
@@ -337,7 +353,6 @@ class OrderBasketShipment extends OrderBasket
 				$params["AMOUNT"] = floatval($item->getQuantity());
 				$params["PRICE"] = $basketItem->getPrice();
 				$params["CURRENCY"] = $basketItem->getCurrency();
-				$params["PRODUCT_PROVIDER_CLASS"] = $basketItem->getProvider();
 				$params["PROPS"] = array();
 
 				/** @var \Bitrix\Sale\BasketPropertyItem $property */
@@ -354,13 +369,15 @@ class OrderBasketShipment extends OrderBasket
 				if(\Bitrix\Main\Loader::includeModule("catalog"))
 				{
 					$productInfo = \CCatalogSku::GetProductInfo($productId);
-					$params["OFFERS_IBLOCK_ID"] = $productInfo["OFFER_IBLOCK_ID"];
-					$params["IBLOCK_ID"] = $productInfo["IBLOCK_ID"];
-					$params["PRODUCT_ID"] = $productInfo["ID"];
+					if ($productInfo)
+					{
+						$params["OFFERS_IBLOCK_ID"] = $productInfo["OFFER_IBLOCK_ID"];
+						$params["IBLOCK_ID"] = $productInfo["IBLOCK_ID"];
+						$params["PRODUCT_ID"] = $productInfo["ID"];
+					}
 				}
 
-				if ($basketItem->isBundleChild())
-					$params["PARENT_BASKET_ID"] = $basketItem->getParentBasketItem()->getId();
+				$params["PARENT_BASKET_ID"] = $basketItem->getParentBasketItemId() ?? 0;
 
 				//If product became bundle, but in saved order it is a simple product.
 				if ($basketItem->getBasketCode() == intval($basketItem->getBasketCode()) && !$basketItem->isBundleParent() && !empty($params['SET_ITEMS']))
@@ -373,6 +390,7 @@ class OrderBasketShipment extends OrderBasket
 			}
 			else
 			{
+				$systemShipmentItem = null;
 				if ($systemShipmentItemCollection)
 				{
 					/** @var \Bitrix\Sale\ShipmentItemCollection $systemShipmentItemCollection */
@@ -532,7 +550,7 @@ class OrderBasketShipment extends OrderBasket
 			static::$jsInited = true;
 		}
 		$data =	$this->prepareData();
-		
+
 		if (!empty($recoveryData))
 		{
 			$data = $this->modifyFromRequest($data, $recoveryData['1']['PRODUCT']);
@@ -611,7 +629,7 @@ class OrderBasketShipment extends OrderBasket
 		\Bitrix\Main\UI\Extension::load('sale.admin_order');
 
 		$data = $this->prepareData();
-		
+
 		return '<script>
 			BX.message({
 				SALE_ORDER_SHIPMENT_VIEW_BASKET_NO_PRODUCTS: "'.\CUtil::JSEscape(Loc::getMessage("SALE_ORDER_SHIPMENT_VIEW_BASKET_NO_PRODUCTS")).'",
@@ -650,16 +668,19 @@ class OrderBasketShipment extends OrderBasket
 			// PREPARE DATA FOR SET_FIELDS
 			foreach ($shipmentBasket as $items)
 			{
-				$items['QUANTITY'] = floatval(str_replace(',', '.', $items['QUANTITY']));
+				$items['QUANTITY'] = floatval(str_replace(',', '.', $items['QUANTITY'] ?? 0));
 				$items['AMOUNT'] = floatval(str_replace(',', '.', $items['AMOUNT']));
 				if (isset($items['BASKET_ID']) && $items['BASKET_ID'] > 0)
 				{
 					if (!$basketItem = $basket->getItemById($items['BASKET_ID']))
 					{
-						$result->addError( new ResultError(
-						   Loc::getMessage('SALE_ORDER_SHIPMENT_BASKET_BASKET_ITEM_NOT_FOUND',  array(
-							   '#BASKET_ITEM_ID#' => $items['BASKET_ID'],
-						   )), 
+						$result->addError(new ResultError(
+							Loc::getMessage(
+								'SALE_ORDER_SHIPMENT_BASKET_BASKET_ITEM_NOT_FOUND',
+								[
+									'#BASKET_ITEM_ID#' => $items['BASKET_ID'],
+								]
+							),
 							'PROVIDER_UNRESERVED_SHIPMENT_ITEM_WRONG_BASKET_ITEM')
 						);
 						return $result;
@@ -672,11 +693,14 @@ class OrderBasketShipment extends OrderBasket
 					$basketCode = $items['BASKET_CODE'];
 					if(!$basketItem = $basket->getItemByBasketCode($basketCode))
 					{
-						$result->addError( new ResultError(
-						   Loc::getMessage('SALE_ORDER_SHIPMENT_BASKET_BASKET_ITEM_NOT_FOUND',  array(
-							   '#BASKET_ITEM_ID#' => $items['BASKET_ID'],
-						   )),
-						   'PROVIDER_UNRESERVED_SHIPMENT_ITEM_WRONG_BASKET_ITEM')
+						$result->addError(new ResultError(
+							Loc::getMessage(
+								'SALE_ORDER_SHIPMENT_BASKET_BASKET_ITEM_NOT_FOUND',
+								[
+									'#BASKET_ITEM_ID#' => $items['BASKET_ID'],
+								]
+							),
+							'PROVIDER_UNRESERVED_SHIPMENT_ITEM_WRONG_BASKET_ITEM')
 						);
 						return $result;
 					}
@@ -685,11 +709,11 @@ class OrderBasketShipment extends OrderBasket
 				$tmp = array(
 					'BASKET_CODE' => $basketCode,
 					'AMOUNT' => $items['AMOUNT'],
-					'ORDER_DELIVERY_BASKET_ID' => $items['ORDER_DELIVERY_BASKET_ID']
+					'ORDER_DELIVERY_BASKET_ID' => $items['ORDER_DELIVERY_BASKET_ID'] ?? null,
 				);
 				$idsFromForm[$basketCode] = array();
 
-				if ($items['BARCODE_INFO'] && (self::$useStoreControl || $basketItem->isSupportedMarkingCode()))
+				if (!empty($items['BARCODE_INFO']) && (self::$useStoreControl || $basketItem->isSupportedMarkingCode()))
 				{
 					foreach ($items['BARCODE_INFO'] as $item)
 					{
@@ -706,7 +730,7 @@ class OrderBasketShipment extends OrderBasket
 						);
 
 						$barcodeCount = 0;
-						if ($item['BARCODE'])
+						if (!empty($item['BARCODE']) && is_array($item['BARCODE']))
 						{
 							foreach ($item['BARCODE'] as $barcode)
 							{
@@ -769,16 +793,18 @@ class OrderBasketShipment extends OrderBasket
 			}
 
 			$shipmentItemStoreCollection = $shipmentItem->getShipmentItemStoreCollection();
-
-			/** @var \Bitrix\Sale\ShipmentItemStore $shipmentItemStore */
-			foreach ($shipmentItemStoreCollection as $shipmentItemStore)
+			if ($shipmentItemStoreCollection)
 			{
-				$shipmentItemId = $shipmentItemStore->getId();
-				if (!isset($idsFromForm[$shipmentItem->getBasketCode()]['BARCODE_IDS'][$shipmentItemId]))
+				/** @var \Bitrix\Sale\ShipmentItemStore $shipmentItemStore */
+				foreach ($shipmentItemStoreCollection as $shipmentItemStore)
 				{
-					$delResult = $shipmentItemStore->delete();
-					if (!$delResult->isSuccess())
-						$result->addErrors($delResult->getErrors());
+					$shipmentItemId = $shipmentItemStore->getId();
+					if (!isset($idsFromForm[$shipmentItem->getBasketCode()]['BARCODE_IDS'][$shipmentItemId]))
+					{
+						$delResult = $shipmentItemStore->delete();
+						if (!$delResult->isSuccess())
+							$result->addErrors($delResult->getErrors());
+					}
 				}
 			}
 		}
@@ -840,34 +866,37 @@ class OrderBasketShipment extends OrderBasket
 
 				/** @var \Bitrix\Sale\ShipmentItemStoreCollection $shipmentItemStoreCollection */
 				$shipmentItemStoreCollection = $shipmentItem->getShipmentItemStoreCollection();
-				if (!$basketItem->isBarcodeMulti() && !$basketItem->isSupportedMarkingCode())
+				if ($shipmentItemStoreCollection)
 				{
-					/** @var Result $r */
-					$r = $shipmentItemStoreCollection->setBarcodeQuantityFromArray($shipmentBasket[$basketItem->getId()]);
-					if(!$r->isSuccess())
+					if (!$basketItem->isBarcodeMulti() && !$basketItem->isSupportedMarkingCode())
 					{
-						$result->addErrors($r->getErrors());
+						/** @var Result $r */
+						$r = $shipmentItemStoreCollection->setBarcodeQuantityFromArray($shipmentBasket[$basketItem->getId()]);
+						if(!$r->isSuccess())
+						{
+							$result->addErrors($r->getErrors());
+						}
 					}
-				}
 
-				if (isset($barcode['ID']) && intval($barcode['ID']) > 0)
-				{
-					/** @var \Bitrix\Sale\ShipmentItemStore $shipmentItemStore */
-					if ($shipmentItemStore = $shipmentItemStoreCollection->getItemById($barcode['ID']))
+					if (isset($barcode['ID']) && intval($barcode['ID']) > 0)
 					{
-						unset($barcode['ID']);
-						$setFieldResult = $shipmentItemStore->setFields($barcode);
+						/** @var \Bitrix\Sale\ShipmentItemStore $shipmentItemStore */
+						if ($shipmentItemStore = $shipmentItemStoreCollection->getItemById($barcode['ID']))
+						{
+							unset($barcode['ID']);
+							$setFieldResult = $shipmentItemStore->setFields($barcode);
 
+							if (!$setFieldResult->isSuccess())
+								$result->addErrors($setFieldResult->getErrors());
+						}
+					}
+					else
+					{
+						$shipmentItemStore = $shipmentItemStoreCollection->createItem($basketItem);
+						$setFieldResult = $shipmentItemStore->setFields($barcode);
 						if (!$setFieldResult->isSuccess())
 							$result->addErrors($setFieldResult->getErrors());
 					}
-				}
-				else
-				{
-					$shipmentItemStore = $shipmentItemStoreCollection->createItem($basketItem);
-					$setFieldResult = $shipmentItemStore->setFields($barcode);
-					if (!$setFieldResult->isSuccess())
-						$result->addErrors($setFieldResult->getErrors());
 				}
 			}
